requirements.txt
# Flask web framework
Flask==2.3.3

# Environment variable management
python-dotenv==1.0.0

# Azure Application Insights integration
opencensus-ext-azure==1.1.15
opencensus-ext-flask==0.8.2

# Azure core libraries (dependencies for opencensus-ext-azure)
azure-core==1.36.0
azure-identity==1.25.1

# Flask dependencies (automatically installed with Flask but listed for clarity)
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# HTTP library for requests
requests==2.32.5

# Other dependencies for Azure integration
msal==1.34.0
msal-extensions==1.3.1
cryptography==46.0.3
PyJWT==2.10.1

# System monitoring (used by opencensus)
psutil==7.1.2

# Core dependencies
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
typing_extensions==4.15.0